chore: move the harness code to its own package
Moving the remaining test data and test CRDS along.
