+++ pkg/test/harness/harness.go
@@ -1,4 +1,4 @@
-package test
+package harness
 
 import (
 	"bytes"
