+++ pkg/test/harness/harness_integration_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package test
+package harness
 
 import (
 	"syscall"
