+++ pkg/test/harness/harness_test.go
@@ -1,4 +1,4 @@
-package test
+package harness
 
 import (
 	"context"
