+++ pkg/test/harness/test_data/cli-test/00-create-pod.yaml
