+++ pkg/test/harness/test_data/cli-test/01-patch.yaml
