+++ pkg/test/harness/test_data/crd-in-step/00-crd.yaml
