+++ pkg/test/harness/test_data/crd-in-step/01-assert.yaml
