+++ pkg/test/harness/test_data/create-or-update/02-update-service.yaml
