+++ pkg/test/harness/test_data/delete-in-step/00-assert.yaml
