+++ pkg/test/harness/test_data/delete-in-step/00-create.yaml
