+++ pkg/test/harness/test_data/delete-in-step/01-assert.yaml
