+++ pkg/test/harness/test_data/delete-in-step/01-create.yaml
