+++ pkg/test/harness/test_data/teststep-apply/00-assert.yaml
