+++ pkg/test/harness/test_data/with-overrides/00-assert.yaml
