+++ pkg/test/harness/test_data/with-overrides/02-directory/assert.yaml
