+++ pkg/test/harness/test_data/with-overrides/03-assert.yaml
